component mapping is fixed
