+++ SQL/CONSTRAINTS_CHECK.sql
@@ -1,0 +1,3 @@
+SELECT *
+  FROM user_cons_columns
+ WHERE constraint_name = '';
