--- src/main/java/com/example/demo/ApplicationRunner.java
@@ -1,13 +1,0 @@
-package com.example.demo;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class ApplicationRunner {
-
-	public static void main(String[] args) {
-		SpringApplication.run(ApplicationRunner.class, args);
-	}
-
-}
