--- src/main/java/com/example/demo/config/SwaggerConfig.java
@@ -1,28 +1,0 @@
-package com.example.demo.config;
-
-import java.util.Collections;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-import springfox.documentation.builders.RequestHandlerSelectors;
-import springfox.documentation.service.ApiInfo;
-import springfox.documentation.service.Contact;
-import springfox.documentation.spi.DocumentationType;
-import springfox.documentation.spring.web.plugins.Docket;
-
-@Configuration
-public class SwaggerConfig {
-
-	private ApiInfo apiInfo() {
-		return new ApiInfo("Resume Builder API", "API for creating and managing resume", "1.0.1", "Terms of Service",
-				new Contact("example.org", "www.example.org", "help@example.org"), "Open-Source",
-				"Open Source License Link", Collections.emptyList());
-	}
-
-	@Bean
-	public Docket api() {
-		return new Docket(DocumentationType.OAS_30).apiInfo(apiInfo()).select().apis(RequestHandlerSelectors.basePackage("com.example.demo"))
-				.build();
-	}
-}
