--- src/main/java/com/example/demo/exception/PersonNotFoundException.java
@@ -1,15 +1,0 @@
-package com.example.demo.exception;
-
-public class PersonNotFoundException extends  RuntimeException{
-    public PersonNotFoundException(String message) {
-        super(message);
-    }
-
-    public PersonNotFoundException(String message, Throwable cause) {
-        super(message, cause);
-    }
-
-    public PersonNotFoundException(Throwable cause) {
-        super(cause);
-    }
-}
