--- src/main/java/com/example/demo/mapping/UrlMapper.java
@@ -1,8 +1,0 @@
-package com.example.demo.mapping;
-
-public class UrlMapper {
-	public final static String globalUrl = "/api/v1/manage/";
-	public final static String dataUrl = "data";
-	public final static String addressUrl = "address";
-
-}
