--- src/main/java/com/example/demo/repository/address/AddressRepository.java
@@ -1,10 +1,0 @@
-package com.example.demo.repository.address;
-
-import com.example.demo.data.Address;
-import org.springframework.data.repository.CrudRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface AddressRepository extends CrudRepository<Address, Long> {
-
-}
