--- src/main/java/com/example/demo/repository/person/PersonRepository.java
@@ -1,10 +1,0 @@
-package com.example.demo.repository.person;
-
-import com.example.demo.data.Person;
-import org.springframework.data.repository.CrudRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface PersonRepository extends CrudRepository<Person, Long> {
-	
-}
