--- src/main/java/com/example/demo/service/person/PersonService.java
@@ -1,17 +1,0 @@
-package com.example.demo.service.person;
-
-import com.example.demo.data.Person;
-import com.example.demo.data.ResponseFormat;
-import java.util.List;
-
-
-public interface PersonService {
-
-    public List<Person> getAllData();
-    public Person getData(Long Id);
-    public ResponseFormat saveData(Person personData);
-    public ResponseFormat alterData(Person personData, Long Id);
-    public ResponseFormat dropAllData();
-    public ResponseFormat dropData(Long Id);
-
-}
