+++ src/main/java/com/example/manager/ApplicationRunner.java
@@ -1,0 +1,13 @@
+package com.example.manager;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class ApplicationRunner {
+
+	public static void main(String[] args) {
+		SpringApplication.run(ApplicationRunner.class, args);
+	}
+
+}
