+++ src/main/java/com/example/manager/data/ResponseFormat.java
@@ -1,0 +1,18 @@
+package com.example.manager.data;
+
+import org.springframework.stereotype.Component;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+@Component
+public class ResponseFormat {
+	
+	private int status;
+	private String message;
+	private long timeStamp;
+}
