+++ src/main/java/com/example/manager/exception/PersonNotFoundException.java
@@ -1,0 +1,15 @@
+package com.example.manager.exception;
+
+public class PersonNotFoundException extends  RuntimeException{
+    public PersonNotFoundException(String message) {
+        super(message);
+    }
+
+    public PersonNotFoundException(String message, Throwable cause) {
+        super(message, cause);
+    }
+
+    public PersonNotFoundException(Throwable cause) {
+        super(cause);
+    }
+}
