+++ src/main/java/com/example/manager/mapping/UrlMapper.java
@@ -1,0 +1,8 @@
+package com.example.manager.mapping;
+
+public class UrlMapper {
+	public final static String globalUrl = "/api/v1/manage/";
+	public final static String dataUrl = "data";
+	public final static String addressUrl = "address";
+
+}
