+++ src/main/java/com/example/manager/repository/address/AddressRepository.java
@@ -1,0 +1,10 @@
+package com.example.manager.repository.address;
+
+import com.example.manager.data.Address;
+import org.springframework.data.repository.CrudRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface AddressRepository extends CrudRepository<Address, Long> {
+
+}
