+++ src/main/java/com/example/manager/repository/person/PersonRepository.java
@@ -1,0 +1,10 @@
+package com.example.manager.repository.person;
+
+import com.example.manager.data.Person;
+import org.springframework.data.repository.CrudRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface PersonRepository extends CrudRepository<Person, Long> {
+	
+}
