+++ src/main/resources/application.properties
@@ -5,6 +5,16 @@ spring.application.name=data-manager
 #logging level
 logging.level.org.springframework=DEBUG
 
+# Database connection settings
+spring.datasource.url=jdbc:oracle:thin:@localhost:1521:ORCLCDB
+spring.datasource.username=C##john
+spring.datasource.password=wick
+
+# Hibernate properties
+spring.jpa.database-platform=org.hibernate.dialect.Oracle12cDialect
+spring.jpa.show-sql=true
+spring.jpa.hibernate.ddl-auto=update
+
 #couple of extra configs
 spring.main.allow-bean-definition-overriding=true
 spring.mvc.pathmatch.matching-strategy=ant_path_matcher
