--- src/test/java/com/example/demo/ApplicationRunnerTests.java
@@ -1,13 +1,0 @@
-package com.example.demo;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ApplicationRunnerTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
