+++ src/test/java/com/example/manager/ApplicationRunnerTests.java
@@ -1,0 +1,13 @@
+package com.example.manager;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class ApplicationRunnerTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
